Use Platform x86 for Windows (it is required) + NuGet
Previously, Windows users set AnyCPU for Visual Studio
but code was actually built x86 (32-bit only). This
can be confusing. Mercurial included is 32-bit only
so the Windows build requires 32-bit which is why the
code is built x86 and with this change it is obvious to the
developer.

Autofac is available as a NuGet package. This change
uses the new convention of each project including a
packages.config that specifies its dependencies and
Autofac is not shipped in the repo.